Move to impl file
