+++ conjure-python-core/src/main/java/com/palantir/conjure/python/poet/EmptySnippet.java
@@ -1,0 +1,39 @@
+/*
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.conjure.python.poet;
+
+import org.immutables.value.Value;
+
+@Value.Immutable
+public interface EmptySnippet extends PythonSnippet {
+
+    @Override
+    @Value.Default
+    default String idForSorting() {
+        return "";
+    }
+
+    @Override
+    default void emit(PythonPoetWriter poetWriter) {
+    }
+
+    class Builder extends ImmutableEmptySnippet.Builder {}
+
+    static Builder builder() {
+        return new Builder();
+    }
+}
