+++ conjure-python-core/src/main/java/com/palantir/conjure/python/processors/packagename/ConstantPackageNameProcessor.java
@@ -1,0 +1,31 @@
+/*
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.conjure.python.processors.packagename;
+
+public final class ConstantPackageNameProcessor implements PackageNameProcessor {
+
+    private final String constantPackageName;
+
+    public ConstantPackageNameProcessor(String constantPackageName) {
+        this.constantPackageName = constantPackageName;
+    }
+
+    @Override
+    public String process(String packageName) {
+        return constantPackageName;
+    }
+}
