+++ conjure-python-core/src/test/resources/services/expected/package_name/__init__.py
@@ -1,7 +1,4 @@
 __all__ = [
-    'another',
-    'product',
-    'product_datasets',
 ]
 
 __conjure_generator_version__ = "0.0.0"
