+++ conjure-python-core/src/test/resources/types/expected/package_name/__init__.py
@@ -1,11 +1,4 @@
 __all__ = [
-    'another',
-    'nested_deeply_nested_service',
-    'nested_service',
-    'nested_service2',
-    'product',
-    'product_datasets',
-    'with_imports',
 ]
 
 __conjure_generator_version__ = "0.0.0"
